Modified assetloader & removed more temporary code
The vectors with all objects in assetloader now keeps track if the model or texture is on GPU or not. This was needed to later make sure that we dont copy the same data(textures for instance) to the GPU more then once. Otherwise this could occur when using the same textures on multiple object in the scene.