Fix for memory leak concerning upgrades.
Made it so that upgradeComponent now gets a new instance of the upgrade instead of a pointer to the upgrade in the map in UpgradeHandler. 

Also added to the function in Player that applies an upgrade if you hit "u". Now it applies rangeTest the first time it is clicked and every other cklick efter that. And it applies meleeTest on the second click and every other click after that.