Optimized the bad solution for text
This solution proved to be about 30% faster than the earlier one. Instead of waiting for the GPU at every removal of a text, it is now filling a buffer until a certain size/limit where it will emptying itself. An higher limit leads to an higher framerate, but also a more unstable one. I found 50 to be a good softspot