New camera
Thanks to Leos changes, the camera now functions on a distance instead of position. Distance is set as a define and can be change accordingly. If an object (Bullet) comes in between the player and the distance of the camera, the new distance is in front of whatever distance the object had.